# frontend/public/openapi.yaml
openapi: 3.1.0
info:
  title: CXDB Gateway API
  version: 0.3.0
  description: Authenticated HTTP access to the CXDB Turn DAG and Blob CAS. Exact turn hydration is required for complete large-trace data.
servers:
  - url: /
security:
  - browserSession: []
  - bearerToken: []
paths:
...
  /api/v1/me:
    get:
      security:
        - browserSession: []
      operationId: currentUser
      responses:
        '200': {description: Current identity and CSRF token}
  /api/v1/tokens:
    get:
      security:
        - browserSession: []
      operationId: listPersonalTokens
      responses:
        '200': {description: Personal token metadata; never plaintext or hashes}
    post:
      security:
        - browserSession: []
      operationId: createPersonalToken
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateToken'}
      responses:
        '201': {description: Token metadata and one-time plaintext}
...
components:
  securitySchemes:
    browserSession: {type: apiKey, in: cookie, name: cxdb_session}
    bearerToken: {type: http, scheme: bearer}
  parameters:
...
    CSRFToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string}}
  schemas:
    CreateToken:
      type: object
      required: [name, scopes]
      properties:
        name: {type: string}
        scopes: {type: array, items: {type: string, enum: ['cxdb:read', 'cxdb:write']}}
        expires_at: {type: string, format: date-time}
...